@@ -18,13 +18,13 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^4.3.2
-  intl: ^0.16.1
+  provider: ^5.0.0
+  intl: ^0.17.0
   http: ^0.12.2
 
   # The following adds the Cupertino Icons font to your application.
