@@ -1,43 +1,46 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_store/providers/cart.dart';
-import 'package:flutter_store/providers/orders.dart';
-import 'package:flutter_store/widgets/cart_item_widget.dart';
 import 'package:provider/provider.dart';
+import 'package:flutter_store/utils/app_string.dart';
+import 'package:flutter_store/providers/cart_provider.dart';
+import 'package:flutter_store/widgets/cart_item_widget.dart';
+import 'package:flutter_store/providers/orders_provider.dart';
 
 class CartPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final Cart cartProvider = Provider.of(context);
+    final CartProvider cartProvider = Provider.of(context);
     final cartItems = cartProvider.items.values.toList();
 
     return Scaffold(
       appBar: AppBar(
-        title: Text('Carrinho'),
+        title: Text(AppString.titleCart),
       ),
       body: Column(
         children: <Widget>[
           Card(
-            margin: EdgeInsets.all(25),
+            margin: EdgeInsets.symmetric(
+              horizontal: 25.0,
+              vertical: 15.0,
+            ),
             child: Padding(
-              padding: EdgeInsets.all(10),
+              padding: EdgeInsets.all(10.0),
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: <Widget>[
                   Text(
-                    'Total',
-                    style: TextStyle(fontSize: 20),
+                    AppString.labelTotal,
+                    style: TextStyle(fontSize: 20.0),
                   ),
                   SizedBox(
-                    width: 20,
+                    width: 20.0,
                   ),
                   Chip(
                     label: Text(
-                      'R\$${cartProvider.totalAmount}',
+                      'R\$${cartProvider.totalAmount.toStringAsFixed(2)}',
                       style: TextStyle(
-                          color: Theme.of(context)
-                              .primaryTextTheme
-                              .headline6
-                              .color),
+                        color:
+                            Theme.of(context).primaryTextTheme.headline6?.color,
+                      ),
                     ),
                     backgroundColor: Theme.of(context).primaryColor,
                   ),
@@ -47,7 +50,7 @@ class CartPage extends StatelessWidget {
               ),
             ),
           ),
-          SizedBox(height: 10),
+          SizedBox(height: 10.0),
           Expanded(
             child: ListView.builder(
               itemCount: cartProvider.itemsCount,
@@ -63,10 +66,10 @@ class CartPage extends StatelessWidget {
 class OrderButtonWidget extends StatefulWidget {
   const OrderButtonWidget({
     Key key,
-    @required this.cartProvider,
+    required this.cartProvider,
   }) : super(key: key);
 
-  final Cart cartProvider;
+  final CartProvider cartProvider;
 
   @override
   _OrderButtonWidgetState createState() => _OrderButtonWidgetState();
@@ -78,6 +81,11 @@ class _OrderButtonWidgetState extends State<OrderButtonWidget> {
   @override
   Widget build(BuildContext context) {
     return TextButton(
+      style: TextButton.styleFrom(
+        textStyle: TextStyle(
+          color: Theme.of(context).primaryColor,
+        ),
+      ),
       onPressed: widget.cartProvider.totalAmount == 0
           ? null
           : () async {
@@ -85,8 +93,10 @@ class _OrderButtonWidgetState extends State<OrderButtonWidget> {
                 _isLoading = true;
               });
 
-              await Provider.of<Orders>(context, listen: false)
-                  .addOrder(widget.cartProvider);
+              await Provider.of<OrderProvider>(
+                context,
+                listen: false,
+              ).addOrder(widget.cartProvider);
 
               setState(() {
                 _isLoading = false;
@@ -94,7 +104,8 @@ class _OrderButtonWidgetState extends State<OrderButtonWidget> {
 
               widget.cartProvider.clear();
             },
-      child: _isLoading ? CircularProgressIndicator() : Text('COMPRAR'),
+      child:
+          _isLoading ? CircularProgressIndicator() : Text(AppString.labelBuy),
     );
   }
 }
