@@ -1,11 +1,15 @@
 import 'package:flutter/material.dart';
 
-class Badge extends StatelessWidget {
+class BadgeWidget extends StatelessWidget {
   final Widget child;
   final String value;
-  final Color color;
+  final Color? color;
 
-  Badge({@required this.child, @required this.value, this.color});
+  BadgeWidget({
+    required this.child,
+    required this.value,
+    this.color,
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -14,23 +18,23 @@ class Badge extends StatelessWidget {
       children: [
         child,
         Positioned(
-          right: 8,
-          top: 8,
+          right: 8.0,
+          top: 8.0,
           child: Container(
-            padding: EdgeInsets.all(2),
+            padding: EdgeInsets.all(2.0),
             decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(10),
-              color: color != null ? color : Theme.of(context).accentColor,
+              borderRadius: BorderRadius.circular(10.0),
+              color: color ?? Theme.of(context).colorScheme.secondary,
             ),
             constraints: BoxConstraints(
-              minHeight: 16,
-              minWidth: 16,
+              minHeight: 16.0,
+              minWidth: 16.0,
             ),
             child: Text(
               value,
               textAlign: TextAlign.center,
               style: TextStyle(
-                fontSize: 10,
+                fontSize: 10.0,
               ),
             ),
           ),
