@@ -1,20 +1,20 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_store/providers/auth.dart';
-import 'package:flutter_store/providers/cart.dart';
 import 'package:flutter_store/utils/app_routes.dart';
 import 'package:flutter_store/utils/app_string.dart';
 import 'package:flutter_store/models/product_model.dart';
+import 'package:flutter_store/providers/cart_provider.dart';
 
 class ProductItemWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final Auth auth = Provider.of(context, listen: false);
-    final Cart cartProvider = Provider.of(context, listen: false);
+    final CartProvider cartProvider = Provider.of(context, listen: false);
     final ProductModel productModel = Provider.of(context, listen: false);
 
     return ClipRRect(
-      borderRadius: BorderRadius.circular(10),
+      borderRadius: BorderRadius.circular(10.0),
       child: GridTile(
         child: GestureDetector(
           onTap: () {
@@ -39,7 +39,7 @@ class ProductItemWidget extends StatelessWidget {
               icon: Icon(productModel.isFavorite
                   ? Icons.favorite
                   : Icons.favorite_border),
-              color: Theme.of(context).accentColor,
+              color: Theme.of(context).colorScheme.secondary,
               onPressed: () => productModel.toggleFavorite(
                 auth.token,
                 auth.userId,
@@ -47,11 +47,11 @@ class ProductItemWidget extends StatelessWidget {
             ),
           ),
           title: Text(
-            productModel.title,
+            productModel.name,
             textAlign: TextAlign.center,
           ),
           trailing: IconButton(
-            color: Theme.of(context).accentColor,
+            color: Theme.of(context).colorScheme.secondary,
             icon: Icon(Icons.add_shopping_cart),
             onPressed: () {
               ScaffoldMessenger.of(context).hideCurrentSnackBar();
