@@ -3,7 +3,7 @@ import 'package:flutter_store/models/product_model.dart';
 final dummyProducts = [
   ProductModel(
     id: 'p1',
-    title: 'Red Shirt',
+    name: 'Red Shirt',
     description: 'A red shirt - it is pretty red!',
     price: 29.99,
     imageUrl:
@@ -11,7 +11,7 @@ final dummyProducts = [
   ),
   ProductModel(
     id: 'p2',
-    title: 'Trousers',
+    name: 'Trousers',
     description: 'A nice pair of trousers.',
     price: 59.99,
     imageUrl:
@@ -19,14 +19,14 @@ final dummyProducts = [
   ),
   ProductModel(
     id: 'p3',
-    title: 'Yellow Scarf',
+    name: 'Yellow Scarf',
     description: 'Warm and cozy - exactly what you need for the winter.',
     price: 19.99,
     imageUrl: 'https://live.staticflickr.com/4043/4438260868_cc79b3369d_z.jpg',
   ),
   ProductModel(
     id: 'p4',
-    title: 'A Pan',
+    name: 'A Pan',
     description: 'Prepare any meal you want.',
     price: 49.99,
     imageUrl:
