@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:flutter_store/providers/cart.dart';
 import 'package:flutter_store/utils/app_string.dart';
+import 'package:flutter_store/models/cart_item_model.dart';
+import 'package:flutter_store/providers/cart_provider.dart';
 
 class CartItemWidget extends StatelessWidget {
   final CartItem cartItem;
 
-  CartItemWidget(this.cartItem);
+  const CartItemWidget(this.cartItem);
 
   @override
   Widget build(BuildContext context) {
     return Dismissible(
-      direction: DismissDirection.endToStart,
       confirmDismiss: (_) {
         return showDialog(
           context: context,
@@ -36,40 +36,45 @@ class CartItemWidget extends StatelessWidget {
         );
       },
       onDismissed: (_) {
-        Provider.of<Cart>(context, listen: false).removeItem(
+        Provider.of<CartProvider>(context, listen: false).removeItem(
           cartItem.productId,
         );
       },
-      key: ValueKey(cartItem.id),
       background: Container(
         color: Theme.of(context).errorColor,
         child: Icon(
           Icons.delete,
           color: Colors.white,
-          size: 40,
+          size: 40.0,
         ),
         alignment: Alignment.centerRight,
-        padding: EdgeInsets.only(right: 20),
-        margin: EdgeInsets.symmetric(horizontal: 15, vertical: 4),
+        padding: const EdgeInsets.only(right: 20.0),
+        margin: const EdgeInsets.symmetric(
+          horizontal: 15.0,
+          vertical: 4.0,
+        ),
       ),
+      key: ValueKey(cartItem.id),
+      direction: DismissDirection.endToStart,
       child: Card(
-        margin: EdgeInsets.symmetric(
-          horizontal: 15,
-          vertical: 4,
+        margin: const EdgeInsets.symmetric(
+          horizontal: 15.0,
+          vertical: 4.0,
         ),
         child: Padding(
-          padding: EdgeInsets.all(8),
+          padding: const EdgeInsets.all(8.0),
           child: ListTile(
             leading: CircleAvatar(
               child: Padding(
-                padding: EdgeInsets.all(5),
+                padding: const EdgeInsets.all(5.0),
                 child: FittedBox(
                   child: Text('${cartItem.price}'),
                 ),
               ),
             ),
             title: Text(cartItem.title),
-            subtitle: Text('Total: R\$ ${cartItem.price * cartItem.quantity}'),
+            subtitle: Text(
+                '${AppString.labelTotal}: R\$ ${cartItem.price * cartItem.quantity}'),
             trailing: Text('${cartItem.quantity}x'),
           ),
         ),
