@@ -1,9 +1,9 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_store/providers/orders.dart';
 import 'package:intl/intl.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_store/models/order_model.dart';
 
 class OrderWidget extends StatefulWidget {
-  final Order order;
+  final OrderModel order;
 
   OrderWidget(this.order);
 
@@ -17,13 +17,14 @@ class _OrderWidgetState extends State<OrderWidget> {
   @override
   Widget build(BuildContext context) {
     return Card(
-      margin: EdgeInsets.all(10.0),
+      margin: const EdgeInsets.all(10.0),
       child: Column(
         children: [
           ListTile(
             title: Text('R\$ ${widget.order.total}'),
-            subtitle:
-                Text(DateFormat('dd/MM/yyyy hh:mm').format(widget.order.date)),
+            subtitle: Text(
+              DateFormat('dd/MM/yyyy hh:mm').format(widget.order.date),
+            ),
             trailing: IconButton(
               icon: Icon(Icons.expand_more),
               onPressed: () {
@@ -35,28 +36,35 @@ class _OrderWidgetState extends State<OrderWidget> {
           ),
           if (_expanded)
             Container(
-              height: (widget.order.products.length * 30.0) + 10,
-              padding: EdgeInsets.symmetric(horizontal: 15.0, vertical: 4.0),
+              height: (widget.order.products.length * 30.0) + 10.0,
+              padding: const EdgeInsets.symmetric(
+                horizontal: 15.0,
+                vertical: 4.0,
+              ),
               child: ListView(
-                children: widget.order.products.map((product) {
-                  return Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: <Widget>[
-                      Text(
-                        product.title,
-                        style: TextStyle(
-                            fontSize: 18.0, fontWeight: FontWeight.bold),
-                      ),
-                      Text(
-                        '${product.quantity} X R\$ ${product.price}',
-                        style: TextStyle(
-                          fontSize: 18.0,
-                          color: Colors.grey,
+                children: widget.order.products.map(
+                  (product) {
+                    return Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: <Widget>[
+                        Text(
+                          product.title,
+                          style: TextStyle(
+                            fontSize: 18.0,
+                            fontWeight: FontWeight.bold,
+                          ),
                         ),
-                      ),
-                    ],
-                  );
-                }).toList(),
+                        Text(
+                          '${product.quantity} X R\$ ${product.price}',
+                          style: TextStyle(
+                            fontSize: 18.0,
+                            color: Colors.grey,
+                          ),
+                        ),
+                      ],
+                    );
+                  },
+                ).toList(),
               ),
             ),
         ],
